[ADD] update profile

+++ src/pages/profile.js
@@ -1,57 +1,71 @@
 import img from "assets/images/cycle.png";
 import axios from "axios";
-import Cookies from "js-cookie";
+import dynamic from "next/dynamic";
 import Image from "next/image";
-import Link from "next/link";
 import { useRouter } from "next/router";
 import { useContext, useEffect, useState } from "react";
-import { useForm } from "react-hook-form";
 import { Store } from "utils/Store";
+import Cookies from "js-cookie";
+import Link from "next/link";
+import { useForm } from "react-hook-form";
 
-const Login = () => {
-  const [user, setUser] = useState();
-  useEffect(() => {
-    const value = localStorage.getItem("userInfo");
-    const user = !!value ? JSON.parse(value) : undefined;
-    setUser(user);
-  }, []);
+function Profile() {
   const {
     handleSubmit,
+    control,
     register,
     formState: { errors },
+    setValue,
   } = useForm();
   const router = useRouter();
   const { redirect } = router.query;
   const { state, dispatch } = useContext(Store);
-  const { userInfo } = state;
+    const { userInfo } = state;
+
+//   const [userInfo, setUserInfo] = useState();
+
+//   useEffect(() => {
+//     const value = localStorage.getItem("userInfo");
+//     const user = !!value ? JSON.parse(value) : undefined;
+//     setUserInfo(user);
+//   }, []);
   useEffect(() => {
-    if (user) {
-      router.push("/");
+    if (!userInfo) {
+      return router.push("/login");
     }
+    setValue("name", userInfo?.name);
+    setValue("email", userInfo?.email);
   }, []);
 
-  const submitHandler = async ({ email, password }) => {
+  const submitHandler = async ({ name, email, password, confirmPassword }) => {
+    if (password !== confirmPassword) {
+      alert("Password don't match");
+      return;
+    }
     try {
-      const { data } = await axios.post("/api/users/login", {
-        email,
-        password,
-      });
-      console.log("login", data);
+      const { data } = await axios.put(
+        "/api/users/profile",
+        {
+          name,
+          email,
+          password,
+        },
+        { headers: { authorization: `Bearer ${userInfo.token}` } }
+      );
+      //   console.log("data", data);
       dispatch({ type: "USER_LOGIN", payload: data });
       localStorage.setItem("userInfo", JSON.stringify(data));
-      // localStorage.setItem("password", password);
-      router.push(redirect || "/");
+      alert("Profile updated successfully");
     } catch (err) {
-      console.log(err);
-      alert(err.response.data ? err.response.data.message : err.message);
+      alert(err.message);
     }
   };
 
   return (
-    <div>
-      <div className="flex items-center justify-center overflow-x-hidden bg-chateau-green-50 lg:overflow-x-auto lg:overflow-hidden">
+    <>
+      <div className="flex items-center justify-center overflow-x-hidden bg-yellow-100 lg:overflow-x-auto lg:overflow-hidden">
         <div className="flex flex-col flex-wrap justify-between w-full border-gray-300 login-container lg:w-4/5 lg:bg-white lg:h-screen lg:border lg:flex-nowrap lg:flex-row group">
-          <div className="relative flex order-2 w-full mt-32 lg:w-1/2 h-28 lg:h-full lg:mt-0 lg:bg-royal-blue lg:order-1">
+          <div className="relative flex order-2 w-full mt-32 lg:w-1/2 h-28 lg:h-full lg:mt-0 lg:bg-theme-yellow-dark lg:order-1">
             <div className="items-center justify-start hidden w-full h-full text-center select-none lg:flex">
               <span className="transform block whitespace-nowrap h-full -rotate-90 text-[55px] 2xl:text-[70px] font-black uppercase text-yellow-300 opacity-0 transition-all group-hover:opacity-100 ml-10 2xl:ml-12 group-hover:-ml-20 2xl:group-hover:ml-26 lg:group-hover:ml-20 duration-1000 lg:duration-700 ease-in-out">
                 Academist
@@ -68,14 +82,40 @@ const Login = () => {
             <div className="hidden w-1/3 ml-auto bg-white lg:block"></div>
           </div>
           <div className="order-1 w-full lg:w-1/2 lg:order-2">
-            <div className="relative flex items-center px-10 pt-16 form-wrapper lg:h-full lg:pt-0">
+            <div className="relative z-10 flex items-center px-10 pt-16 form-wrapper lg:h-full lg:pt-0">
               <div className="w-full space-y-2">
                 <div className="flex items-end justify-center mb-8 space-x-3 text-center form-caption">
-                  <span className="text-3xl font-semibold text-royal-blue center">
-                    Log in now
+                  <span className="text-3xl font-semibold text-royal-blue">
+                    Account Settings
                   </span>
                 </div>
                 <form onSubmit={handleSubmit(submitHandler)}>
+                  <div className="form-element">
+                    <label className="space-y-0.5 w-full lg:w-4/5 block mx-auto">
+                      <span className="block text-lg tracking-wide text-gray-800">
+                        Name
+                      </span>
+                      <span className="block">
+                        <input
+                          type="text"
+                          name="name"
+                          // eslint-disable-next-line react/jsx-props-no-spreading
+                          {...register("name", {
+                            required: {
+                              value: true,
+                              message: "You most enter name",
+                            },
+                          })}
+                          className={`block w-full px-4 py-3 placeholder-gray-500 border-gray-300 rounded-md shadow focus:ring-blue-500 focus:border-blue-500 focus:outline-none focus:ring-2
+               ${errors.name ? "ring-2 ring-red-500" : null}`}
+                          placeholder="Full name"
+                        />
+                        <span className="py-2 text-sm text-red-400">
+                          {errors?.name?.message}
+                        </span>
+                      </span>
+                    </label>
+                  </div>
                   <div className="form-element">
                     <label className="space-y-0.5 w-full lg:w-4/5 block mx-auto">
                       <span className="block text-lg tracking-wide text-gray-800">
@@ -135,7 +175,7 @@ const Login = () => {
                           })}
                           className={`block w-full px-4 py-3 placeholder-gray-500 border-gray-300 rounded-md shadow focus:ring-blue-500 focus:border-blue-500 focus:outline-none focus:ring-2
                ${errors.password ? "ring-2 ring-red-500" : null}`}
-                          placeholder="password"
+                          placeholder="Password"
                         />
                         <span className="py-2 text-sm text-red-400">
                           {errors?.password?.message}
@@ -144,45 +184,63 @@ const Login = () => {
                     </label>
                   </div>
                   <div className="form-element">
-                    <div className="flex items-center justify-between w-full py-2 mx-auto lg:w-4/5">
+                    <label className="space-y-0.5 w-full lg:w-4/5 block mx-auto">
+                      <span className="block text-lg tracking-wide text-gray-800">
+                        Conform Password
+                      </span>
+                      <span className="block">
+                        <input
+                          type="password"
+                          name="confirmPassword"
+                          // eslint-disable-next-line react/jsx-props-no-spreading
+                          {...register("confirmPassword", {
+                            required: {
+                              value: true,
+                              message: "You most enter confirm Password",
+                            },
+                            minLength: {
+                              value: 6,
+                              message: "confirm Password lenth is more then 5",
+                            },
+                          })}
+                          className={`block w-full px-4 py-3 placeholder-gray-500 border-gray-300 rounded-md shadow focus:ring-blue-500 focus:border-blue-500 focus:outline-none focus:ring-2
+               ${errors.confirmPassword ? "ring-2 ring-red-500" : null}`}
+                          placeholder="Confirm Password"
+                        />
+                        <span className="py-2 text-sm text-red-400">
+                          {errors?.confirmPassword?.message}
+                        </span>
+                      </span>
+                    </label>
+                  </div>
+                  <div className="form-element">
+                    <div className="flex items-center py-2 mx-auto lg:w-4/5">
                       <label className="flex items-center space-x-2 tracking-wide text-gray-800 select-none">
                         <input type="checkbox" name="" id="" />
                         <span className="block tracking-wide text-gray-800">
                           Remember me
                         </span>
                       </label>
-                      <Link
-                        href="#"
-                        className="block tracking-wide text-gray-800 border-b border-gray-300"
-                      >
-                        Forgot Password?
-                      </Link>
                     </div>
                   </div>
 
                   <div className="form-element">
-                    <span className="block w-full mx-auto lg:w-full ">
+                    <span className="block w-full mx-auto lg:w-4/5 ">
                       <input
                         type="submit"
-                        className="flex w-full px-6 py-3 mx-auto text-lg text-white bg-indigo-600 border-0 rounded cursor-pointer lg:w-4/5 focus:outline-none hover:bg-aquamarine-800"
-                        value="Log in now"
+                        className="flex w-full px-6 py-3 text-lg text-white bg-indigo-600 border-0 rounded cursor-pointer focus:outline-none hover:bg-aquamarine-800"
+                        value="Update Account"
                       />
                     </span>
                   </div>
                 </form>
-                <p className="text-center d-block">
-                  Don't have an account?{" "}
-                  <Link href="/register">
-                    <a className="text-royal-blue">Create an account</a>
-                  </Link>
-                </p>
               </div>
             </div>
           </div>
         </div>
       </div>
-    </div>
+    </>
   );
-};
+}
 
-export default Login;
+export default dynamic(() => Promise.resolve(Profile), { ssr: false });
